requirements.txt
Flask==2.2.5
requests==2.31.0
beautifulsoup4==4.12.2
PyPDF2==3.0.1
Pillow==10.0.0
python-magic-bin==0.4.14
dnspython==2.4.1
python-dotenv==1.0.0